Move ConfigurationChecker from server-demo to server-core
